+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning since version 1.0.0.
 - Change top nav breadcrumb colours for section pages in DHP theme
 - Rotate certs for login.gov staging
 - Remove "local" cert: only 2 envs exist now
+- Update style map for "w:" and "heading 8" styles
 
 ### Fixed